+++ Lunar/Comment View/CreateCommentPopover.swift
@@ -13,10 +13,10 @@ struct CreateCommentPopover: View {
   @Default(.activeAccount) var activeAccount
 
   @EnvironmentObject var commentsFetcher: CommentsFetcher
-  @Environment(\.dismiss) var dismiss
 
   @State private var userInputComment: String = ""
   @State private var userInputCommentEntry: String = ""
+  @Binding var showingCreateCommentPopover: Bool
 
   var post: RealmPost
   var parentID: Int?
@@ -102,7 +102,7 @@ struct CreateCommentPopover: View {
           .bold()
         Spacer()
         Button {
-          dismiss()
+          showingCreateCommentPopover = false
         } label: {
           Image(systemSymbol: .xmarkCircleFill)
             .font(.largeTitle)
@@ -139,7 +139,7 @@ struct CreateCommentPopover: View {
             if response == "success" {
               commentsFetcher.loadContent(isRefreshing: true)
               notificationHaptics.notificationOccurred(.success)
-              dismiss()
+              showingCreateCommentPopover = false
             } else {
               notificationHaptics.notificationOccurred(.error)
               print("ERROR SENDING COMMENT")