+++ Lunar/Inbox Tab/InboxView.swift
@@ -23,32 +23,23 @@ struct InboxView: View {
   var body: some View {
     NavigationView {
       List {
-        if messages.count == 0 {
-          Text("No messages found")
+        if messages.isEmpty || activeAccount.actorID.isEmpty {
+          Text(activeAccount.actorID.isEmpty ? "Login to view messages" : "No messages found")
             .listRowBackground(Color.clear)
+            .foregroundStyle(.secondary)
         } else {
           Section {
             ForEach(messages, id: \.messageID) { message in
-              if message.creatorActorID != activeAccount.actorID {
-                MessageItem(message: message)
-              }
+              MessageItem(message: message)
             }
-          } header: {
-            Text("Received")
           }
-          Section {
-            ForEach(messages, id: \.messageID) { message in
-              if message.creatorActorID == activeAccount.actorID {
-                MessageItem(message: message)
-              }
-            }
-          } header: {
-            Text("Sent")
-          }
+          .listRowBackground(Color.clear)
+          .listRowSeparator(.hidden)
         }
       }
-      .listStyle(.insetGrouped)
+      .listStyle(.plain)
       .navigationTitle("Inbox")
+      .background(Color("postListBackground"))
       .onAppear {
         guard !privateMessagesRetrieved else { return }
         reloadMessages()
@@ -88,6 +79,8 @@ enum messageDirection {
 struct MessageItem: View {
   @Default(.fontSize) var fontSize
   @Default(.activeAccount) var activeAccount
+  @Default(.accentColor) var accentColor
+  @Default(.accentColorString) var accentColorString
 
   @ObservedRealmObject var message: RealmPrivateMessage
 
@@ -95,9 +88,9 @@ struct MessageItem: View {
     message.creatorActorID == activeAccount.actorID ? .outgoing : .incoming
   }
 
-//  var fromTo: String {
-//    messageDirection == .incoming ? "from" : "to"
-//  }
+  //  var fromTo: String {
+  //    messageDirection == .incoming ? "from" : "to"
+  //  }
 
   var name: String {
     messageDirection == .incoming ? message.creatorName : message.recipientName
@@ -111,29 +104,43 @@ struct MessageItem: View {
     messageDirection == .incoming ? message.creatorAvatar : message.recipientAvatar
   }
 
-  var body: some View {
-    VStack(alignment: .leading) {
-      HStack(alignment: .center) {
-//        fromToLabel
-        fromAvatar
-        header
-        Spacer()
-      }
-      messageBody
-      timeAgoHeader
+  var chatBubbleDirection: BubblePosition {
+    messageDirection == .incoming ? .left : .right
+  }
+
+  var chatBubbleColor: Color {
+    if messageDirection == .incoming {
+      let color = accentColorString == "Default" ? Color.messageBubbleBackgroundBlue : .accentColor
+      return color
+    } else {
+      return Color.messageBubbleBackgroundGray
     }
   }
 
-//
-//  var fromToLabel: some View {
-//    Text()
-//      .foregroundStyle(.secondary)
+//  var composerSettings: some View {
+//    Image(systemSymbol: accentColorString == "Default" ? .textBubbleFill : .textBubble)
+//      .symbolRenderingMode(accentColorString == "Default" ? .hierarchical : .monochrome)
+//      .foregroundStyle(accentColorString == "Default" ? .gray : accentColor)
 //  }
+
+  var body: some View {
+    ChatBubble(position: chatBubbleDirection, color: chatBubbleColor) {
+      VStack(alignment: .leading) {
+        HStack(alignment: .center) {
+          fromAvatar
+          header
+          Spacer()
+        }
+        messageBody
+        timeAgoHeader
+      }
+    }
+  }
 
   var header: some View {
     HStack(alignment: .firstTextBaseline, spacing: 1) {
       Text("\(name)")
-        .textCase(.uppercase)
+        .bold()
         .fixedSize()
 
       Text("@\(URLParser.extractDomain(from: actorID))")
@@ -164,55 +171,21 @@ struct MessageItem: View {
       }
     }
     .processors([.resize(width: 50)])
-
     .frame(width: 20, height: 20)
     .transition(.opacity)
   }
 
   var messageBody: some View {
-    Markdown { message.messageContent }
-      .markdownTextStyle(\.text) { FontSize(fontSize) }
-      .markdownTheme(.gitHub)
-      .padding(.bottom, 2)
+    Text(message.messageContent)
+      .font(.system(size: 15))
+      .padding(.bottom, 3)
   }
 }
 
-//  @ViewBuilder
-//  var actionButtons: some View {
-//    ReactionButton(
-//      icon: SFSafeSymbols.SFSymbol.checkmarkCircleFill,
-//      color: Color.blue,
-//      active: $message.messageRead.not,
-//      opposite: .constant(false)
-//    )
-//    .hapticFeedbackOnTap(style: .rigid)
-//
-//    ReactionButton(
-//      icon: SFSafeSymbols.SFSymbol.flagCircleFill,
-//      color: Color.red,
-//      active: $message.messageRead.not,
-//      opposite: .constant(false)
-//    )
-//    .hapticFeedbackOnTap(style: .rigid)
-//
-//    ReactionButton(
-//      icon: SFSafeSymbols.SFSymbol.arrowshapeTurnUpLeftCircleFill,
-//      color: Color.orange,
-//      active: $message.messageRead.not,
-//      opposite: .constant(false)
-//    )
-//    .hapticFeedbackOnTap(style: .rigid)
-//
-//    ReactionButton(
-//      icon: SFSafeSymbols.SFSymbol.docCircleFill,
-//      color: Color.gray,
-//      active: $message.messageRead.not,
-//      opposite: .constant(false)
-//    )
-//    .hapticFeedbackOnTap(style: .rigid)
-//  }
-
 #Preview {
-  MessageItem(message: MockData().privateMessage)
-    .previewLayout(.sizeThatFits)
+  VStack {
+    MessageItem(message: MockData().privateMessageIncoming)
+    MessageItem(message: MockData().privateMessageIncoming)
+  }
+  .previewLayout(.sizeThatFits)
 }